Rename delete helper and improve docstrings
