+++ helpers/delete_selected_tracks.py
@@ -2,7 +2,12 @@ import bpy
 
 
 def delete_selected_tracks():
-    """Delete selected tracks if a clip is active."""
+    """Delete selected tracks if a clip is active.
+
+    This helper is called by various operators such as
+    :class:`~operators.tracking.cleanup.CLIP_OT_track_cleanup` and
+    :class:`~operators.tracking.detect.CLIP_OT_detect_button`.
+    """
     clip = getattr(bpy.context.space_data, "clip", None)
     if clip is None:
         return False
